feat: add logic to delete and update experience

+++ src/modules/user/experience/experience.controller.ts
@@ -2,29 +2,37 @@ import { Body, Controller, Delete, Get, Patch, Post, Request, UseGuards } from '
 import { ExperienceService } from './experience.service';
 import { CreateExperienceDto } from './dto/create-experience.dto';
 import { AuthGuard } from 'src/modules/auth/auth.guard';
+import { RemoveExperienceDto } from './dto/remove-experience.dto';
+import { UpdateExperienceDto } from './dto/update-experience.dto';
 
 
+@UseGuards(AuthGuard)
 @Controller('user/experience')
 export class ExperienceController {
   constructor(
     private experienceService: ExperienceService
   ) { }
 
-  @UseGuards(AuthGuard)
   @Post('add')
   async addExperience(
     @Body() createExperienceDto: CreateExperienceDto,
-    @Request() req: {userId: string}
+    @Request() req: { userId: string }
   ) {
-    return this.experienceService.addExperience({...createExperienceDto, userId: req.userId});
+    return this.experienceService.addExperience({ ...createExperienceDto, userId: req.userId });
   }
   @Patch('update')
-  async updateExperience() {
-    return this.experienceService.updateExperience();
+  async updateExperience(
+    @Body() updateExperienceDto: UpdateExperienceDto,
+    @Request() req: { userId: string }
+  ) {
+    return this.experienceService.updateExperience({ ...updateExperienceDto, userId: req.userId });
   }
   @Delete('delete')
-  async removeExperience() {
-    return this.experienceService.removeExperience();
+  async removeExperience(
+    @Body() removeExperienceDto: RemoveExperienceDto,
+    @Request() req: { userId: string }
+  ) {
+    return this.experienceService.removeExperience({ ...removeExperienceDto, userId: req.userId });
   }
 
 }
